Turn System Delegates
Make game State controllers to handle turns and phases.
Moved Enums to a separate file.
Made Base widget and hud classes for the UI Systems.
Split the fire system to a new class in anticipation of having to do multithreading later.
Updated the Build file to include the slate targets.